- Added inappwebview - Added paymob service - Added paymob repo - Added payment gateway view - Added address model - Added foodie user model - Added Payment Gateway Args - Added button gradient disabled - Added new fonts (font20primarycolorsemibold - fon24secondarycolorbold) - Added loading indicator in the custom elevated button - Added error model - Added api result - Added network exceptions - Added recepit model - Added payment cubit - Added success recepit alert dialog

+++ lib/features/cart/logic/cart_cubit/cart_cubit.dart
@@ -9,7 +9,7 @@ part 'cart_state.dart';
 class CartCubit extends Cubit<CartState> {
   CartCubit() : super(const CartState.emptyCart());
   List<FoodItem> cartItems = [];
-  int price = 0;
+  double price = 0;
 
   void addItemToCart(FoodItem item) {
     final newFoodItem = item.copyWith();
@@ -24,7 +24,7 @@ class CartCubit extends Cubit<CartState> {
   }
 
   void updateCheckoutPrice() {
-    int currentPrice = 0;
+    double currentPrice = 0;
     for (var item in cartItems) {
       currentPrice += (item.totalPrice * item.quantity);
     }
